Rename message_schemas to json_schemas
It's only to do with JSON schemas rather than schemas for all message
types so change the name to be more specific.
